Add `runuser` package for run as www-data user
Users very often user need to run php scripts in shell manually, but now shell works as `root` user, so PHP scripts work as root too. And if they generates some files, those files got `root` owner.
As I understand, we can't switch shell to `www-data` by default, because of using `s6-overlay`.
So workaround is to add `sudo` or `runuser` tools to run php scripts as `www-data` user.
I prefer `runuser` because it is more lightweight, so I add it to list of installed packages.
With this we can run scripts like this:
```
runuser -u  www-data -- ./scripts/some-script-that-creates-files.php
```